check users role

+++ app/Http/Controllers/API/BlogController.php
@@ -11,7 +11,6 @@ use Illuminate\Support\Facades\Auth;
 
 class BlogController extends BaseController
 {
-
     public function index()
     {
         $auth_user = Auth::user('api');
@@ -20,7 +19,7 @@ class BlogController extends BaseController
         } else if($auth_user->hasRole('user')) {
             $blogs = $auth_user->blogs;
         }
-        return $this->sendResponse(BlogResource::collection($blogs), 'Posts fetched.');
+        return $this->sendResponse(BlogResource::collection($blogs), 'Blogs fetched.');
     }
 
     public function store(Request $request)
@@ -31,10 +30,11 @@ class BlogController extends BaseController
             'description' => 'required'
         ]);
         if ($validator->fails()) {
-            return $this->sendError($validator->errors());
+            return $this->sendError($validator->errors(), null, 422);
         }
-        $blog = Blog::create($input);
-        return $this->sendResponse(new BlogResource($blog), 'Post created.');
+        $input['user_id'] = Auth::user('api')->id;
+        $blog = Blog::create($input);   
+        return $this->sendResponse(new BlogResource($blog), 'Blog created.');
     }
 
 
@@ -42,13 +42,33 @@ class BlogController extends BaseController
     {
         $blog = Blog::find($id);
         if (is_null($blog)) {
-            return $this->sendError('Post does not exist.');
+            return $this->sendError('Blog does not exist.');
         }
-        return $this->sendResponse(new BlogResource($blog), 'Post fetched.');
+
+        $auth_user = Auth::user('api');
+        if($auth_user->hasRole('user')) {
+            if($blog->user_id !== $auth_user->id) {
+                return $this->sendError("You are not authorized to access this Blog.", null, 401);
+            }
+        }
+        
+        return $this->sendResponse(new BlogResource($blog), 'Blog fetched.');
     }
 
-    public function update(Request $request, Blog $blog)
+    public function update(Request $request, $id)
     {
+        $blog = Blog::find($id);
+        if (is_null($blog)) {
+            return $this->sendError('Blog does not exist.');
+        }
+
+        $auth_user = Auth::user('api');
+        if($auth_user->hasRole('user')) {
+            if($blog->user_id !== $auth_user->id) {
+                return $this->sendError("You are not authorized to update this Blog.", null, 401);
+            }
+        }
+        
         $input = $request->all();
 
         $validator = Validator::make($input, [
@@ -57,19 +77,32 @@ class BlogController extends BaseController
         ]);
 
         if ($validator->fails()) {
-            return $this->sendError($validator->errors());
+            return $this->sendError($validator->errors(), null, 422);
         }
 
         $blog->title = $input['title'];
         $blog->description = $input['description'];
+        $blog->user_id = $auth_user->id;
         $blog->save();
 
-        return $this->sendResponse(new BlogResource($blog), 'Post updated.');
+        return $this->sendResponse(new BlogResource($blog), 'Blog updated.');
     }
 
-    public function destroy(Blog $blog)
+    public function destroy($id)
     {
+        $blog = Blog::find($id);
+        if (is_null($blog)) {
+            return $this->sendError('Blog does not exist.');
+        }
+
+        $auth_user = Auth::user('api');
+        if($auth_user->hasRole('user')) {
+            if($blog->user_id !== $auth_user->id) {
+                return $this->sendError("You are not authorized to delete this Blog.", null, 401);
+            }
+        }
+
         $blog->delete();
-        return $this->sendResponse([], 'Post deleted.');
+        return $this->sendResponse([], 'Blog deleted.');
     }
 }
